+++ __mocks__/react-native-dotenv.js
@@ -1,0 +1,4 @@
+module.exports = {
+    EXPO_PUBLIC_OPENAI_PROXY_URL: "",
+    EXPO_PUBLIC_OPENAI_PROXY_KEY: "",
+};